fix: removed logic from microapp component
