+++ packages/user-preferences/src/sandbox-user-preferences-repo.ts
@@ -1,7 +1,7 @@
 import { UserPreferencesRepo } from './user-preferences-repo';
 import { SandboxPreferencesOptions, type UserPreferencesData } from './types';
 
-export class SandboxUserPreferences implements UserPreferencesRepo {
+export class SandboxUserPreferencesRepo implements UserPreferencesRepo {
   #mockPreferences: UserPreferencesData;
   #listeners: Set<(data?: UserPreferencesData) => void> = new Set();
